Extract network tab
